// src/app/_restOrdenes/model/detalleOrden.ts
/**
 * Servicio Ordenes Kallsonys
 * Servicio para la gestión de ordenes en  OMS Kallsonys
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Detalle de las ordenes de la plataforma Kallsonys
 */
export interface DetalleOrden { 
    idDetOrden?: number;
    idOrden?: number;
    idProducto?: number;
    cantidad?: number;
    valorUnidad?: number;
    idProveedor?: number;
    estado?: string;
}